Rename Quiz Storage Test to Quiz Storage Test - Max
This code saves answer from quiz to a dictionary then saves this dictionary as a "binary .txt" which can be opened later as a list. to see its true functionality take the quiz save some answers don't load them. then restart the programme press n twice and type load_quiz() this will load the dictionary from the previous time the programme was open.

I will update this to work in GUI by utilising a save screen that will enable the user to save and load the previous values in the various entry box's through out the programme. I will also attempt to write some code using the navigation test to enable the user to open up the screen they were on prior to closing.